Refactored folder and package names, removed main func and added tests.
