+++ strings/levenshteindistance/levenshteinDistance.go
@@ -4,16 +4,11 @@ Parameters: two strings to compare and weights of insertion, substitution and de
 Output: distance between both strings
 */
 
-package main
-
-import (
-	"fmt"
-)
+package levenshteindistance
 
 func levenshteinDistance(str1, str2 string, icost, scost, dcost int) int {
 	row1 := make([]int, len(str2)+1)
 	row2 := make([]int, len(str2)+1)
-	var tmp []int
 
 	for i := 1; i <= len(str2); i++ {
 		row1[i] = i * icost
@@ -39,35 +34,8 @@ func levenshteinDistance(str1, str2 string, icost, scost, dcost int) int {
 				}
 			}
 		}
-
-		tmp = row1
-		row1 = row2
-		row2 = tmp
+		row1, row2 = row2, row1
 	}
 
 	return row1[len(row1)-1]
-}
-
-func min(str1, str2 int) int {
-	if str1 < str2 {
-		return str1
-	}
-	return str2
-}
-
-func main() {
-	str1 := "stingy"
-	str2 := "ring"
-
-	// Using weight 1 for insertion, substitution and deletion
-	strDistance1 := levenshteinDistance(str1, str2, 1, 1, 1)
-	fmt.Printf("Distance between \"%s\" and \"%s\" is: %d.\n", str1, str2, strDistance1)
-	// Output: Distance between "stingy" and "ring" is: 3.
-
-	// Using weight 1 for insertion/substitution and weight 3 for deletion
-	strDistance2 := levenshteinDistance(str1, str2, 1, 1, 3)
-	fmt.Printf("Distance between \"%s\" and \"%s\" is: %d.\n", str1, str2, strDistance2)
-	// Output: Distance between "stingy" and "ring" is: 7.
-
-	return
 }
